Log when local build really starts, to better explain the build process
Without this, stopping & removing containers seemed to take a very long
time, as it included all the time to upload the build context & start
the build itself. This is now more explicit, which should help people
realise that delays might be due to a large build context locally, and
generally make it clearer what's going on.

Change-Type: patch