Fix https://github.com/resin-io/resin-cli/issues/824 by ensuring local push source is absolute
Not 100% clear why this works... Seems like odd behaviour in pkg.
When we tar the directory, if we use '.', we see a 'snapshot' dir
there, which we can't actually read. Pkg creates a snapshot dir
that the app is run out of, though it shouldn't really be appearing
here. Regardless - using absolute paths all the time works identically
and solves the issue.

Change-Type: patch